vid 4

+++ ToDoList/Views/ListView.swift
@@ -12,19 +12,32 @@ struct ListView: View {
     @EnvironmentObject var vm: ListViewModel
     
     var body: some View {
-        List {
-            ForEach(vm.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear(duration: 0.2)) {
-                            vm.updateItem(item: item)
-                        }
+        VStack {
+            if vm.items.isEmpty {
+                Spacer()
+                ContentUnavailableView(
+                    "You're all done",
+                    systemImage: "checkmark",
+                    description: Text("Tap Add button if you have to-dos")
+                )
+                Spacer()
+            } else {
+                List {
+                    ForEach(vm.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear(duration: 0.2)) {
+                                    vm.updateItem(item: item)
+                                }
+                            }
                     }
+                    .onDelete(perform: vm.deleteItem)
+                    .onMove(perform: vm.moveItem)
+                }
+                .padding(.top)
+                .listStyle(PlainListStyle())
             }
-            .onDelete(perform: vm.deleteItem)
-            .onMove(perform: vm.moveItem)
         }
-        .listStyle(PlainListStyle())
         .navigationTitle("To-Do List 📝")
         .toolbar {
             ToolbarItem(placement: .topBarLeading) {
